requirements.txt
google-genai>=1.61.0
python-dotenv>=1.0.0
